pyproject.toml
[tool.poetry]
name = "library-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.95.1"}
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.13"
python-dotenv = "^1.0.0"
black = "^23.3.0"
pytest = "^7.3.1"
alembic = "^1.11.1"
faker = "^18.9.0"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
